+++ tests/src/cy_trace.c
@@ -21,12 +21,12 @@ void cy_trace(cy_t* const         cy, // cppcheck-suppress constParameterPointer
     static const int32_t mega = 1000000;
     static const int32_t kilo = 1000;
     (void)fprintf(stderr, //
-                  "CY_TRACE(%p) %06lld.%03lld %s:%04u:%s: ",
+                  "CY_TRACE(%p) %06jd.%03jd %s:%04ju:%s: ",
                   (void*)cy,
-                  (long long)(now / mega),
-                  (long long)((now % mega) / kilo),
+                  (intmax_t)(now / mega),
+                  (intmax_t)((now % mega) / kilo),
                   file_name,
-                  (unsigned)line,
+                  (uintmax_t)line,
                   func);
 
     // Print the message.
